backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "HWDC 2025 FastAPI Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "agno",
    "ollama>=0.5.4",
    "openai>=1.108.1",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio>=0.23",
    "httpx",
    "ruff",
    "mypy",
    "pre-commit>=4.3.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W", "C4", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
files = ["src/"]

[tool.pytest.ini_options]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "api: marks tests related to the API layer",
    "application: marks tests related to the application layer",
]
